Fix #349 Fix pegsless mode in DOSXYZnrc source 20
If you are using a BEAM shared library geometry with
source 20, it is assumed that the pegs data for the BEAM
geometry is the same as that for DOSXYZnrc.  Previously,
if you ran DOSXYZnrc in pegsless mode, this information was
not correctly passed on to the BEAM simulation.  Now it
is, and the BEAM simulation also runs in pegsless mode.
(Don't forget to define the media in the BEAM input file!)